Use stable property information during preview.
And support property paths in stables.

The value of each stable property is copied from the old state of the
resource. If only a subset of a value is stable, its other elements
(either properties or array elements) are left unknown. This ensures
that we have appropriate shapes for nested property values.