Invert order of applying transformations
The goal of transformations is to allow outer code to override the behaviour of inner code.  So to be consistent, we should apply parent transformations after applying child transformations, from the bottom of the hierarchy up to the stack.  This ensure stack level transformations have the ability to observe and override if necessary all other layers of transformations.