Fix bugs in the original change.
The fundamental issue is that we had code that assumed that any output
where `isKnown` resolves to `false` would also resolve to a value that
contained some first-class unknown. When this assumption was violated,
it was possible to end up with outputs with values that were incorrectly
treated as known, especially when used with the new variant of `apply`.