Add ability to opt-in to using yarn instead of npm
This change adds support for setting `PULUMI_PREFER_YARN` to true to
opt-in to preferring yarn over npm. If `PULUMI_PREFER_YARN` is truthy,
but yarn cannot be found on $PATH, we fallback to using npm. If npm
can't be found on $PATH, we provide a more helpful error message.