Restore the work to add Publish release builds of Pulumi (#3875)
Fixes: #3874

As part of #3656, the build.proj was changed around to support
building and running our tests on windows. The work in that PR added
a lot to get Python, Go and DotNet included in our master and PR runs
for those tests

This PR restores the work to allow the original work for non-PRs (i.e.
tagged builds) to work along side this. It adds a `Release` target which
targets the same build targets as before. It is wrapped in a condition that
will mean it can run when AppVeyor is dealing with a tag (i.e a release)

This means that the existing work to allow all of the master and pr builds
to continue running all of our integration tests will still work as
expected