Improve the import failure error message
If we fail to import a resource due to a state mismatch, the
error today is fairly generic:

    error: inputs to import do not match the existing resource

Although the information is available in the normal diff -- both
alongside the resource in the tree view and in the "details" view
during the preview -- the error itself is initially jarring in that
it doesn't tell you much about why the failure occurred.

This change adds a little bit more information to the error
message, including the ID of the resource and all unique
keys that were found to differ. The result is a bit nicer:

    error: imported resource sg-04aeda9a214730248's property
        'ingress' does not match the existing value

This fixes #3737.