Emit policy config validation errors as diagnostic messages
Instead of returning these as a single combined error, emit these as individual diagnostic messages. This allows these errors to be seen in the Pulumi Console. Also, make it so that all validation errors for all policy packs are returned.

We should still consider adding a new diagnostic event for these, which would allow improved grouping and display, but that requires changes to both the CLI and service and can be done subsequently.

Also, cleaned up some things and added more tests.