Resurrect pkg/util/archive
When a user runs `pulumi policy publish` for a Python policy pack, we need to package up a directory of code and send it to the service. This commit brings back some existing code we had for doing this for PPCs, as it was before it was deleted (originally at pkg/util/archive). Subsequent commits will make the necessary changes to make use of this for Python policy packs.