Support secret default values in schema and codegen
Currently, the secretness is only applied to default values used as input values for resources.  It is not applied to the defaults expoed in the `config` modules.  Doing the latter may also be desired - but (a) it would be a breaking change as the types would need to change from primitives to Outputs and (b) users can always manually wrap value retreived from this module with `pulumi.secret` if desired (a choice they do not have for the default resource inputs case).

Part of pulumi/pulumi-azure#576.