improve dev version detection
Currently the `isDevVersion` function only tests for "dev" versions.
When installing pulumi from Homebrew with the `--HEAD` flag we get an
alpha version locally. This can mean the user gets the upgrade prompt
when it's not valid.

This change changes the logic to check against a regex for all the
possible semver `pre` versions.

It also adds a unit test to ensure we're getting the right result